moving vectori to linmath
